requirements.txt
absl-py==2.1.0
addict==2.4.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
configargparse==1.7
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
descartes==1.1.0
docker-pycreds==0.4.0
easydict==1.9
einops==0.8.0
entmax==1.3
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.13.1
fire==0.6.0
flake8==7.1.0
flask==3.0.3
flask-cors==5.0.0
flatbuffers==24.3.25
fsspec==2024.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
gitpython==3.1.43
google-auth==2.31.0
google-auth-oauthlib==1.0.0
grpcio==1.64.1
hf-xet==1.1.4
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.6
imageio==2.34.2
importlib-metadata==8.0.0
iniconfig==2.0.0
iopath==0.1.10
ipython==8.12.3
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.11
llvmlite==0.36.0
lyft-dataset-sdk==0.0.8
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mmengine==0.9.1
mmsegmentation==1.0.0
mpmath==1.3.0
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.2
numba==0.53.0
numpy==1.24.4
nuscenes-devkit==1.1.11
oauthlib==3.2.2
onnx==1.16.0
onnx-graphsurgeon==0.5.2
onnxruntime==1.19.2
onnxruntime-gpu==1.19.2
open3d==0.18.0
opencv-python==4.10.0.84
packaging==24.0
panda==0.3.1
pandas==2.0.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pkgutil-resolve-name==1.3.10
plotly==5.22.0
pluggy==1.5.0
plyfile==1.0.3
portalocker==2.10.1
prettytable==3.11.0
prompt-toolkit==3.0.47
protobuf==5.28.1
psutil==6.0.0
ptflops==0.7.2.2
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycocotools==2.0.7
pycodestyle==2.12.0
pyflakes==3.2.0
pygments==2.18.0
pyquaternion==0.9.9
pytest==8.2.2
pytz==2024.1
pywavelets==1.4.1
pyyaml==6.0.1
referencing==0.35.1
regex==2024.11.6
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
sentencepiece==0.1.97
sentry-sdk==2.14.0
setproctitle==1.3.3
shapely==1.8.5.post1
smmap==5.0.1
stack-data==0.6.3
sympy==1.13.2
tabulate==0.9.0
tenacity==8.4.2
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardx==2.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1-2209072238
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.6.13
tokenizers==0.20.3
torch==1.8.2+cu111
torchaudio==0.8.2
torchvision==0.9.2+cu111
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.0
trimesh==2.35.39
tzdata==2024.1
ultralytics==8.3.158
ultralytics-thop==2.0.14
urllib3==2.2.1
wandb==0.18.0
wcwidth==0.2.13
werkzeug==3.0.3
widgetsnbextension==4.0.11
yacs==0.1.8
yapf==0.40.2
mmcv==2.0.0rc4
mmdet==3.0.0rc6
mmdet3d==1.1.0rc0



